+++ states_rebuilder_package/lib/src/states_rebuilder.dart
@@ -37,11 +37,12 @@ class StatesRebuilder implements Subject {
   ///key holds the observer tags and the value holds the observers
   ///_observers = {"tag" : [observer1, observer2, ...]}
   ///Observers are  automatically add and removed by [StateBuilder] in the [State.initState] and [State.dispose]  methods.
-  final LinkedHashMap<String, List<ObserverOfStatesRebuilder>> _observersMap =
-      LinkedHashMap<String, List<ObserverOfStatesRebuilder>>();
+  final LinkedHashMap<String, Set<ObserverOfStatesRebuilder>> _observersMap =
+      LinkedHashMap<String, Set<ObserverOfStatesRebuilder>>();
+  Set<ObserverOfStatesRebuilder> _observersSet = <ObserverOfStatesRebuilder>{};
 
   /// observers getter
-  Map<String, List<ObserverOfStatesRebuilder>> observers() => _observersMap;
+  Map<String, Set<ObserverOfStatesRebuilder>> observers() => _observersMap;
   bool get hasObservers => _observersMap.isNotEmpty;
 
   ///Holds user defined void callback to be executed after removing all observers.
@@ -52,17 +53,14 @@ class StatesRebuilder implements Subject {
   void addObserver({ObserverOfStatesRebuilder observer, String tag}) {
     assert(observer != null);
     assert(tag != null);
-    Map<String, List<ObserverOfStatesRebuilder>> observersTemp = {};
-    observersTemp.addAll(_observersMap);
-    _observersMap.clear();
-    //observers are add at the beginning of the list.
-    if (observersTemp[tag] == null) {
-      _observersMap[tag] = <ObserverOfStatesRebuilder>[observer];
-      _observersMap.addAll(observersTemp);
+
+    if (_observersMap[tag] == null) {
+      _observersMap[tag] = <ObserverOfStatesRebuilder>{observer};
     } else {
-      observersTemp[tag] = [observer, ...observersTemp[tag]];
-      _observersMap.addAll(observersTemp);
+      _observersMap[tag] = {observer, ..._observersMap[tag]};
     }
+
+    _observersSet = {observer, ..._observersSet};
   }
 
   @override
@@ -89,6 +87,7 @@ class StatesRebuilder implements Subject {
     );
 
     _observersMap[tag].remove(observer);
+    _observersSet.remove(observer);
     if (_observersMap[tag].isEmpty) {
       _observersMap.remove(tag);
       if (_observersMap.isEmpty) {
@@ -102,6 +101,14 @@ class StatesRebuilder implements Subject {
       }
     }
   }
+
+  // dynamic _tag;
+  // bool _isExclusive = false;
+  // StatesRebuilder tag(dynamic tag, [bool isExclusive = false]) {
+  //   _tag = tag;
+  //   _isExclusive = isExclusive;
+  //   return this;
+  // }
 
   /// You call [rebuildState] inside any of your logic classes that extends [StatesRebuilder].
   ///
@@ -146,21 +153,18 @@ class StatesRebuilder implements Subject {
     bool isOnSetStateCalledOrNull = onSetState == null;
 
     if (tags == null) {
-      _observersMap.forEach(
-        (tag, observers) {
-          for (ObserverOfStatesRebuilder observer in observers.reversed) {
-            observer.update(
-              isOnSetStateCalledOrNull
-                  ? null
-                  : (context) {
-                      isOnSetStateCalledOrNull = true;
-                      onSetState(context);
-                    },
-              this is ReactiveModel ? this : null,
-            );
-          }
-        },
-      );
+      for (ObserverOfStatesRebuilder observer in _observersSet) {
+        observer.update(
+          isOnSetStateCalledOrNull
+              ? null
+              : (context) {
+                  isOnSetStateCalledOrNull = true;
+                  onSetState(context);
+                },
+          this is ReactiveModel ? this : null,
+        );
+      }
+
       return;
     }
 
@@ -200,5 +204,6 @@ class StatesRebuilder implements Subject {
 class StatesRebuilderInternal {
   static addAllToObserverMap(StatesRebuilder from, StatesRebuilder to) {
     to?._observersMap?.addAll(from._observersMap);
+    to?._observersSet?.addAll(from._observersSet);
   }
 }